% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseMethods.R
\name{mean}
\alias{mean}
\alias{mean,sparse_numeric-method}
\title{Sparse mean}
\usage{
mean(x)

\S4method{mean}{sparse_numeric}(x)
}
\arguments{
\item{x}{A 'sparse_numeric' object}
}
\value{
A 'sparse_numeric' object
}
\description{
Computers the mean of a full vector represetned by a sparse object
}
\section{Functions}{
\itemize{
\item \code{mean(sparse_numeric)}: Mean of a sparse vector

}}
