% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseMethods.R
\name{norm}
\alias{norm}
\alias{norm,sparse_numeric-method}
\title{Sparse norm}
\usage{
norm(x)

\S4method{norm}{sparse_numeric}(x)
}
\arguments{
\item{x}{A 'sparse_numeric' object}
}
\value{
A 'sparse_numeric' object
}
\description{
Computes the norm of a sparse vector
}
\section{Functions}{
\itemize{
\item \code{norm(sparse_numeric)}: Normalize a sparse vector

}}
