% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseMethods.R
\name{plot,sparse_numeric,sparse_numeric-method}
\alias{plot,sparse_numeric,sparse_numeric-method}
\title{Plot Two Sparse Numeric Vectors}
\usage{
\S4method{plot}{sparse_numeric,sparse_numeric}(x, y, ...)
}
\arguments{
\item{x}{A \code{sparse_numeric} object}

\item{y}{A \code{sparse_numeric} object}

\item{...}{Additional graphical parameters}
}
\description{
Plots the nonzero values of two sparse vectors on the same graph
}
