% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseMethods.R
\name{sparse_add}
\alias{sparse_add}
\alias{sparse_add,sparse_numeric,sparse_numeric-method}
\alias{+,sparse_numeric,sparse_numeric-method}
\title{Sparse addition}
\usage{
sparse_add(x, y)

\S4method{sparse_add}{sparse_numeric,sparse_numeric}(x, y)

\S4method{+}{sparse_numeric,sparse_numeric}(e1, e2)
}
\arguments{
\item{x}{A 'sparse_numeric' object}

\item{y}{A 'sparse_numeric' object}

\item{e1}{First operand (sparse_numeric)}

\item{e2}{Second operand (sparse_numeric)}
}
\value{
A new 'sparse_numeric' object
}
\description{
Adds two sparse numeric vectors
}
\section{Functions}{
\itemize{
\item \code{sparse_add(x = sparse_numeric, y = sparse_numeric)}: Add two sparse vectors

\item \code{e1 + e2}: '+' symbol for sparse_numeric vectors

}}
