% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseMethods.R
\name{sparse_crossprod}
\alias{sparse_crossprod}
\alias{sparse_crossprod,sparse_numeric,sparse_numeric-method}
\title{Sparse cross production}
\usage{
sparse_crossprod(x, y)

\S4method{sparse_crossprod}{sparse_numeric,sparse_numeric}(x, y)
}
\arguments{
\item{x}{A 'sparse_numeric' object}

\item{y}{A 'sparse_numeric' object}
}
\value{
A 'sparse_numeric' object
}
\description{
Computes the dot product of two sparse vectors
}
\section{Functions}{
\itemize{
\item \code{sparse_crossprod(x = sparse_numeric, y = sparse_numeric)}: Cross products two sparse vectors

}}
