% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseMethods.R
\name{sparse_mult}
\alias{sparse_mult}
\alias{sparse_mult,sparse_numeric,sparse_numeric-method}
\alias{*,sparse_numeric,sparse_numeric-method}
\title{Sparse multiplication}
\usage{
sparse_mult(x, y)

\S4method{sparse_mult}{sparse_numeric,sparse_numeric}(x, y)

\S4method{*}{sparse_numeric,sparse_numeric}(e1, e2)
}
\arguments{
\item{x}{A 'sparse_numeric' object}

\item{y}{A 'sparse_numeric' object}

\item{e1}{First operand (sparse_numeric)}

\item{e2}{Second operand (sparse_numeric)}
}
\value{
A 'sparse_numeric' object
}
\description{
Multiplication of two sparse numeric vectors
}
\section{Functions}{
\itemize{
\item \code{sparse_mult(x = sparse_numeric, y = sparse_numeric)}: Multiples two sparse vectors

\item \code{e1 * e2}: '*' symbol for sparse_numeric vectors

}}
