% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseMethods.R
\docType{class}
\name{sparse_numeric-class}
\alias{sparse_numeric-class}
\title{Sparse Numeric Vector Class}
\description{
Defining an S4 class for storing numeric vectors by
holding only nonzero values and their positions
}
\section{Slots}{

\describe{
\item{\code{value}}{A numeric vector of nonzero values}

\item{\code{pos}}{Integer vector giving positions of each nonzero value}

\item{\code{length}}{Integer specifying length of the vector}
}}

