% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseMethods.R
\name{sparse_sub}
\alias{sparse_sub}
\alias{sparse_sub,sparse_numeric,sparse_numeric-method}
\alias{-,sparse_numeric,sparse_numeric-method}
\title{Sparse subtraction}
\usage{
sparse_sub(x, y)

\S4method{sparse_sub}{sparse_numeric,sparse_numeric}(x, y)

\S4method{-}{sparse_numeric,sparse_numeric}(e1, e2)
}
\arguments{
\item{x}{A 'sparse_numeric' object}

\item{y}{A 'sparse_numeric' object}

\item{e1}{First operand (sparse_numeric)}

\item{e2}{Second operand (sparse_numeric)}
}
\value{
A 'sparse_numeric' object
}
\description{
Subtracts two sparse vectors
}
\section{Functions}{
\itemize{
\item \code{sparse_sub(x = sparse_numeric, y = sparse_numeric)}: Subtracts two sparse vectors

\item \code{e1 - e2}: '-' symbol for sparse_numeric vectors

}}
