% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparseMethods.R
\name{standardize}
\alias{standardize}
\alias{standardize,sparse_numeric-method}
\title{Sparse standardize}
\usage{
standardize(x)

\S4method{standardize}{sparse_numeric}(x)
}
\arguments{
\item{x}{A 'sparse_numeric' object}
}
\value{
A 'sparse_numeric' object
}
\description{
Standardizes a sparse vector
}
\section{Functions}{
\itemize{
\item \code{standardize(sparse_numeric)}: Standardize a sparse vector

}}
